Get rid of Int in the debug info tests and preform some general cleanup.
<rdar://problem/20856296> Extinguish 'Int' from DebugInfo tests

Swift SVN r31177

+++ test/DebugInfo/ASTSection.swift
@@ -22,7 +22,7 @@
 // A type that should be serialized.
 class Foo {
 init() { }
-func bar() -> Int { return 42 }
+func bar() -> Int64 { return 42 }
 }
 
 // Some toplevel code that should not be serialized.
