+++ test/DebugInfo/argument.swift
@@ -3,7 +3,7 @@
 func markUsed<T>(t: T) {}
 
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_arg_variable, name: "arg", arg: 1,{{.*}} line: [[@LINE+1]]
-func a(arg : Int)
+func a(arg : Int64)
 {
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_auto_variable, name: "local",{{.*}} line: [[@LINE+1]]
   var local = arg
@@ -12,25 +12,25 @@ func a(arg : Int)
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_arg_variable, name: "a", arg: 1,{{.*}} line: [[@LINE+3]]
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_arg_variable, name: "b", arg: 2,{{.*}} line: [[@LINE+2]]
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_arg_variable, name: "c", arg: 3,{{.*}} line: [[@LINE+1]]
-func many(a: Int, b: (Int, Int), c: Int) -> Int {
+func many(a: Int64, b: (Int64, Int64), c: Int64) -> Int64 {
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_auto_variable, name: "i1",{{.*}} line: [[@LINE+1]]
   var i1 = a
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_auto_variable, name: "i2",{{.*}} line: [[@LINE+2]]
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_auto_variable, name: "i3",{{.*}} line: [[@LINE+1]]
-  var (i2, i3) : (Int, Int) = b
+  var (i2, i3) : (Int64, Int64) = b
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_auto_variable, name: "i4",{{.*}} line: [[@LINE+1]]
   var i4 = c
   return i1+i2+i3+i4
 }
 
 class A {
-  var member : Int
+  var member : Int64
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_arg_variable, name: "a", arg: 1,{{.*}} line: [[@LINE+1]]
-  init(a: Int) { member = a }
+  init(a: Int64) { member = a }
 
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_arg_variable, name: "offset", arg: 1,{{.*}} line: [[@LINE+2]]
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_arg_variable, name: "self", arg: 2,{{.*}} line: [[@LINE+1]]
-  func getValuePlus(offset: Int) -> Int {
+  func getValuePlus(offset: Int64) -> Int64 {
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_auto_variable, name: "a",{{.*}} line: [[@LINE+1]]
     var a = member
     return a+offset
@@ -39,7 +39,7 @@ class A {
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_arg_variable, name: "factor", arg: 1,{{.*}} line: [[@LINE+3]]
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_arg_variable, name: "offset", arg: 2,{{.*}} line: [[@LINE+2]]
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_arg_variable, name: "self", arg: 3,{{.*}} line: [[@LINE+1]]
-  func getValueTimesPlus(factor: Int, offset: Int) -> Int {
+  func getValueTimesPlus(factor: Int64, offset: Int64) -> Int64 {
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_auto_variable, name: "a",{{.*}} line: [[@LINE+1]]
     var a = member
 // CHECK-DAG: !DILocalVariable(tag: DW_TAG_auto_variable, name: "f",{{.*}} line: [[@LINE+1]]
@@ -57,12 +57,12 @@ class A {
 // Curried functions have their arguments backwards.
 // CHECK: !DILocalVariable(tag: DW_TAG_arg_variable, name: "b", arg: 1,{{.*}} line: [[@LINE+2]]
 // CHECK: !DILocalVariable(tag: DW_TAG_arg_variable, name: "a", arg: 2,{{.*}} line: [[@LINE+1]]
-func uncurry (a: Int) (b: Int) -> (Int, Int) {
+func uncurry (a: Int64) (b: Int64) -> (Int64, Int64) {
   return (a, b)
 }
 
 // CHECK: !DILocalVariable(tag: DW_TAG_arg_variable, name: "x", arg: 1,{{.*}} line: [[@LINE+2]]
 // CHECK: !DILocalVariable(tag: DW_TAG_arg_variable, name: "y", arg: 2,{{.*}} line: [[@LINE+1]]
-func tuple(x: Int, y: (Int, Float, String)) -> Int {
+func tuple(x: Int64, y: (Int64, Float, String)) -> Int64 {
   return x+y.0;
 }
