+++ test/DebugInfo/return.swift
@@ -1,15 +1,15 @@
 // RUN: %target-swift-frontend %s -g -emit-ir -o - | FileCheck %s
 
 class X {
-  init (i : Int) { x = i }
-  var x : Int
+  init (i : Int64) { x = i }
+  var x : Int64
 }
 
 // CHECK: define {{.*}}ifelseexpr
-public func ifelseexpr() -> Int {
+public func ifelseexpr() -> Int64 {
   var x = X(i:0); 
   // CHECK: [[META:%.*]] = call %swift.type* @_TMaC6return1X()
-  // CHECK: [[X:%.*]] = call %C6return1X* @_TFC6return1XCfMS0_FT1iSi_S0_(i{{32|64}} 0, %swift.type* [[META]])
+  // CHECK: [[X:%.*]] = call %C6return1X* @_TFC6return1XCfMS0_FT1iVSs5Int64_S0_(i64 0, %swift.type* [[META]])
   // CHECK:  @swift_release to void (%C6return1X*)*)(%C6return1X* [[X]])
   if true {
     x.x++;
